requirements.txt
tensorflow==2.13.0
fastapi==0.103.2
uvicorn==0.23.2
streamlit==1.27.2
pillow==10.0.1
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.1
matplotlib==3.8.0
seaborn==0.13.0
locust==2.17.0
python-multipart==0.0.6
requests==2.31.0
h5py==3.9.0
opencv-python-headless==4.8.1.78
tqdm==4.66.1
nbformat==5.9.2
ipykernel==6.25.2
